add new route for /settings
