+++ src/app.html
@@ -3,7 +3,7 @@
 	<head>
 		<meta charset="utf-8" />
 		<link rel="icon" href="%sveltekit.assets%/favicon.png" />
-		<link rel="stylesheet" href="fonts/inter.css">
+		<link rel="stylesheet" href="/fonts/inter.css">
 		<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1" />
 		%sveltekit.head%
 	</head>
